requirements.txt
Django==4.0.6
psycopg2==2.8.6
uWSGI==2.0.19.1
django-import-export==2.8.0